@@ -1,8 +1,8 @@
 apply plugin: 'com.android.application'
+apply plugin: 'com.google.gms.google-services'
 
 android {
-    compileSdkVersion 25
-    buildToolsVersion "24.0.2"
+    compileSdkVersion 31
 
     defaultConfig {
         applicationId "news.androidtv.tvapprepo"
