@@ -2,14 +2,11 @@ package news.androidtv.tvapprepo.activities;
 
 import android.app.Activity;
 import android.content.ComponentName;
-import android.content.DialogInterface;
-import android.content.Intent;
 import android.content.pm.ResolveInfo;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v7.app.AlertDialog;
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AlertDialog;
 import android.util.Log;
-import android.view.ContextThemeWrapper;
 import android.view.Gravity;
 import android.view.View;
 import android.view.ViewGroup;
