@@ -17,10 +17,10 @@ package news.androidtv.tvapprepo.presenters;
 import android.content.Context;
 import android.graphics.Bitmap;
 import android.graphics.drawable.Drawable;
-import android.support.v17.leanback.widget.ImageCardView;
-import android.support.v17.leanback.widget.Presenter;
-import android.support.v7.graphics.Palette;
-import android.transition.Transition;
+import androidx.leanback.widget.ImageCardView;
+import androidx.leanback.widget.Presenter;
+import androidx.palette.graphics.Palette;
+
 import android.util.Log;
 import android.view.ViewGroup;
 
