@@ -1,7 +1,7 @@
 package news.androidtv.tvapprepo.presenters;
 
-import android.support.v17.leanback.widget.ImageCardView;
-import android.support.v17.leanback.widget.Presenter;
+import androidx.leanback.widget.ImageCardView;
+import androidx.leanback.widget.Presenter;
 import android.view.ContextThemeWrapper;
 import android.view.ViewGroup;
 import android.widget.ImageView;
