@@ -1,8 +1,8 @@
 package news.androidtv.tvapprepo.ui;
 
 import android.app.Activity;
-import android.content.Intent;
-import android.support.annotation.NonNull;
+
+import androidx.annotation.NonNull;
 import android.util.Log;
 import android.view.ContextThemeWrapper;
 import android.widget.EditText;
